# swagger/ swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InFrame API",
    "description": "InFrame API",
    "version": "0.1.9"
  },
  "servers": [
    {
      "url": "http://api.example.com/v1",
      "description": "Optional server description, e.g. Main (production) server"
    },
    {
      "url": "http://staging-api.example.com",
      "description": "Optional server description, e.g. Internal staging server for testing"
    }
  ],
  "paths": {
    "/login": {
     "post":{
        "tags":[
          "login"
        ],"parameters": [
          {
          "in": "body",
            "name": "body",
            "description": "Register",
            "schema": {
              "$ref": "#/definitions/login"
              } 
          }
      ],
      "responses": {
          "201": {
              "description": "Register Success"
          },
          "404": {
              "description": "Failed. Bad Request"
          }
      }  
      }
    },"/signup": {
      "post":{
        "tags":[
          "signup"
        ],"parameters": [
          {
          "in": "body",
            "name": "body",
            "description": "Register",
            "schema": {
              "$ref": "#/definitions/signup"
              } 
          }
      ],
      "responses": {
          "201": {
              "description": "Register Success"
          },
          "404": {
              "description": "Failed. Bad Request"
          }
      }  
      }
    },"/signupCode/:email": {
      "post":{
        "tags":[
          "signupCode"
        ],"parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Register",
            "schema": {
              "$ref": "#/definitions/code"
              } 
          }
      ],
      "responses": {
          "201": {
              "description": "Register Success"
          },
          "404": {
              "description": "Failed. Bad Request"
          }
      }  

      }
    },"/findPassword": {
      "post":{
        "tags":[
          "findPassword"
        ],"parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Register",
            "schema": {
              "$ref": "#/definitions/email"
              } 
          }
      ],
      "responses": {
          "201": {
              "description": "Register Success"
          },
          "404": {
              "description": "Failed. Bad Request"
          }
      }  

      }
      },"/passwordCode/:email": {
          "post":{
            "tags":[
              "passwordCode"
            ],"parameters": [
              {
                "in": "body",
                "name": "body",
                "description": "Register",
                "schema": {
                  "$ref": "#/definitions/code"
                  } 
              }
          ],
          "responses": {
              "201": {
                  "description": "Register Success"
              },
              "404": {
                  "description": "Failed. Bad Request"
              }
          }  
    
          }
          },"/newPassword/:email": {
              "post":{
                "tags":[
                  "newPassword"
                ],"parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "Register",
                    "schema": {
                      "$ref": "#/definitions/password"
                      } 
                  }
              ],
              "responses": {
                  "201": {
                      "description": "Register Success"
                  },
                  "404": {
                      "description": "Failed. Bad Request"
                  }
              }  
        
              }
              }
  },
  "definitions":{
    "user": {
      "properties": {
        "email" : {
          "type": "string"
        },
        "password" : {
          "type": "string"
        },
        "code" : {
          "type": "string"
        }
      }
    },"mail": {
      "properties": {
        "code" : {
          "type": "string"
        }
      }
    },"login": {
      "properties": {
        "email" : {
          "type": "string"
        },
        "password" : {
          "type": "string"
        }
      }
    },
    "signup": {
      "properties": {
        "email" : {
          "type": "string"
        },
        "password" : {
          "type": "string"
        }
      }
    },
    "code": {
      "properties": {
        "code" : {
          "type": "string"
        }
      }
    },
    "email": {
      "properties": {
        "email" : {
          "type": "string"
        }
      }
    },
    "password": {
      "properties": {
        "password" : {
          "type": "string"
        }
      }
    }
  }
}